RUBY-32905 Handle stable releases with negative patch number correctly
There was an assumption that patch level can be negative only for rc/preview ruby versions. But that is not generally true. It doesn't work with ruby 3.4.0 which is stable and has patch version -1. For such case it breaks mechanism of guessing right dir with core sources.

Instead of looking for negative patch version in core sources generation process, we need to check ruby version for `-rc`/`-preview` suffix. If ruby version doesn't have such prefix we need to add patch version (even negative) to the name of the dir with core sources for this version. In that way mechanism of guessing dirs works fine.